Move some functions into services/repository
